requirements.txt
chardet==5.2.0
Nuitka==2.6.4
ordered-set==4.1.0
pillow==10.4.0
pycryptodome==3.21.0
PyMySQL==1.1.1
pypdfium2==4.30.1
pypinyin==0.53.0
PySide6==6.1.3
pywin32==308
reportlab==4.3.0
shiboken6==6.1.3
zstandard==0.23.0
